refactor: isolate application composition roots
