+++ gui/tests/app.smoke.test.ts
@@ -5,72 +5,13 @@
 import { afterEach, describe, expect, it, vi } from 'vitest';
 import { cleanup, render, screen, fireEvent, within } from '@testing-library/svelte';
 import App from '../src/App.svelte';
-import { Client } from '../src/lib/ipc/client';
+import { createAppCtx } from '../src/lib/app';
 import { DemoCoreTransport } from '../src/lib/dev/democore';
 import type { AppCtx } from '../src/lib/ctx';
 import { readBoot } from '../src/lib/ipc/boot';
-import { ConnectionStore } from '../src/lib/stores/connection.svelte';
-import { ThemeStore } from '../src/lib/stores/theme.svelte';
-import { UiStore } from '../src/lib/stores/ui.svelte';
-import { PlaybackStore } from '../src/lib/stores/playback.svelte';
-import { QueueStore } from '../src/lib/stores/queue.svelte';
-import { SearchStore } from '../src/lib/stores/search.svelte';
-import { LibraryStore } from '../src/lib/stores/library.svelte';
-import { AiStore } from '../src/lib/stores/ai.svelte';
-import { DownloadsStore } from '../src/lib/stores/downloads.svelte';
-import { PlaylistsStore } from '../src/lib/stores/playlists.svelte';
-import { TransferStore } from '../src/lib/stores/transfer.svelte';
-import { AccountsStore } from '../src/lib/stores/accounts.svelte';
-import { SettingsStore } from '../src/lib/stores/settings.svelte';
-import { AnimStore } from '../src/lib/stores/anim.svelte';
-import { KeymapStore } from '../src/lib/stores/keymap.svelte';
-import { LyricsStore } from '../src/lib/stores/lyrics.svelte';
-import { WhyGemStore } from '../src/lib/stores/whygem.svelte';
-import { ToastStore } from '../src/lib/stores/toasts.svelte';
 
 function assemble(): AppCtx {
-  const client = new Client(new DemoCoreTransport());
-  const connection = new ConnectionStore(client);
-  const toasts = new ToastStore();
-  toasts.attach(client);
-  const ctx: AppCtx = {
-    boot: readBoot(),
-    client,
-    demo: true,
-    connection,
-    theme: new ThemeStore(client),
-    ui: new UiStore(),
-    playback: new PlaybackStore(client),
-    queue: new QueueStore(client, (message) => toasts.show('error', message)),
-    search: new SearchStore(client),
-    library: new LibraryStore(client),
-    ai: new AiStore(client),
-    downloads: new DownloadsStore(client),
-    playlists: new PlaylistsStore(client),
-    transfer: new TransferStore(client),
-    accounts: new AccountsStore(client),
-    settings: new SettingsStore(client, (message) => toasts.show('error', message)),
-    anim: new AnimStore(client),
-    keymap: new KeymapStore(client),
-    lyrics: new LyricsStore(client),
-    whygem: new WhyGemStore(client),
-    toasts,
-  };
-  client.sub([
-    'player',
-    'queue',
-    'lyrics',
-    'search',
-    'library',
-    'playlists',
-    'ai',
-    'downloads',
-    'transfer',
-    'accounts',
-    'settings',
-    'system',
-  ]);
-  return ctx;
+  return createAppCtx(readBoot(), new DemoCoreTransport());
 }
 
 const settle = () => new Promise((r) => setTimeout(r, 250));
